% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{gss1972}
\alias{gss1972}
\title{The General Social Survey 1972}
\format{
A data frame with 1,613 rows and 4 variables
}
\usage{
data(gss1972)
}
\description{
The general Social Survey (GSS) 1972 data.
The variables are as follows:
}
\details{
\itemize{
  \item age Age of respondents.
  \item occ10 Census occupation code (2010).
  \item occ Census occupation code (1970).
  \item birthyr Birth cohort of respondents.
}
}
\keyword{datasets}
