% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{gss1989}
\alias{gss1989}
\title{The General Social Survey 1989}
\format{
A data frame with 1,537 rows and 5 variables
}
\usage{
data(gss1989)
}
\description{
The general Social Survey (GSS) 1989 data.
}
\details{
The variables are as follows:

\itemize{
  \item age Age of respondents.
  \item occ10 Census occupation code (2010).
  \item occ Census occupation code (1970).
  \item occ80 Census occupation code (1980).
  \item birthyr Birth cohort of respondents.
}
}
\keyword{datasets}
