% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{gss2018}
\alias{gss2018}
\title{The General Social Survey 2018}
\format{
A data frame with 2,348 rows and 3 variables
}
\usage{
data(gss2018)
}
\description{
The general Social Survey (GSS) 2018 data.
}
\details{
The variables are as follows:

\itemize{
  \item age Age of respondents.
  \item occ10 Census occupation code (2010).
  \item birthyr Birth cohort of respondents.
}
}
\keyword{datasets}
