% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{input_yearocc}
\alias{input_yearocc}
\title{Occupation Code and Occupation Titles}
\format{
A data frame with 2,917 rows and 3 variables. This data is used for
internal functions in
extracting occupational titles.
}
\usage{
data(input_yearocc)
}
\description{
A data set containing occupation code, title, and the year the occupational
code was coded.
}
\details{
\itemize{
  \item occ Occupational codes.
  \item occ_desc Occupational titles.
  \item input The year occupations are coded (e.g., 1950, 1960, 1970, 1980,
  1990, 2000, 2010).
}
}
\keyword{datasets}
