% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{ipums_data_occ1950_rank}
\alias{ipums_data_occ1950_rank}
\title{Occupational Percentile Rank for Census Occupations (1950)}
\format{
A data frame with 5,033 rows and 4 variables
}
\usage{
data(ipums_data_occ1950_rank)
}
\description{
Occupational percentile rank for census occupations in 1950.
}
\details{
The variables are as follows:

\itemize{
  \item cohort Birth cohort
  \item occupation Occupation titles
  \item OCC1950 Census occupation code (1950)
  \item prank Calculated percentile rank
}
}
\keyword{datasets}
