% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{ipums_data_occ1990_rank}
\alias{ipums_data_occ1990_rank}
\title{Occupational Percentile Rank for Census Occupations (1990)}
\format{
A data frame with 3,940 rows and 4 variables
}
\usage{
data(ipums_data_occ1990_rank)
}
\description{
Occupational percentile rank for census occupations in 1990.
The variables are as follows:
}
\details{
\itemize{
  \item OCC1990 Census occupation code (1990)
  \item Title Occupation titles
  \item cohort Birth cohort
  \item prank Calculated percentile rank
}
}
\keyword{datasets}
