% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{ipums_data_occ2010_rank}
\alias{ipums_data_occ2010_rank}
\title{Occupational Percentile Rank for Census Occupations (2010)}
\format{
A data frame with 4,650 rows and 13 variables
}
\usage{
data(ipums_data_occ2010_rank)
}
\description{
Occupational percentile rank for census occupations in 2010.
The variables are as follows:
}
\details{
\itemize{
  \item OCC2010 Census occupation code (2010)
  \item Title Occupation titles
  \item cohort Birth cohort
  \item prank Calculated percentile rank
}
}
\keyword{datasets}
