% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_rank.R
\name{occ_rank}
\alias{occ_rank}
\title{Constructing an occupation-based socioeconomic index that can easily incorporate occupational structure changes
with the method proposed by Song and Xie (2020).}
\usage{
occ_rank(occ, cohort, input_year, output_year, export = "prank")
}
\arguments{
\item{occ}{occupation code in the input year}

\item{cohort}{birth cohort}

\item{input_year}{input year of occupational code(s), e.g., 1950, 1960, 1970, 1980, 1990, 2000, 2010.}

\item{output_year}{The year occupational code(s) is/are crosswalked to, can only be 1950, 1990, and 2010}

\item{export}{the format of returned results, "prank" or "table". If NULL, the default is "prank",
a vector of occupation percentile rank(s) for input occupation(s). Otherwise, "table" returns a data
frame including input year, input occupational code(s), output year, output occupational code(s), and
percentile rank(s) for the specified occupation(s).}
}
\value{
occupational index or a data frame including all detailed information.
}
\description{
Constructing an occupation-based socioeconomic index that can easily incorporate occupational structure changes
with the method proposed by Song and Xie (2020).
}
\examples{
library(OccRank)
library(haven)
occ_rank(occ = 2, cohort = 1970, input_year = 1970, output_year = 1950)
occ_rank(occ = 2, cohort = 1970, input_year = 1970,
         output_year = 1950,export="table")
occ_rank(occ = 2, cohort = 1970, input_year = 1970,
         output_year = 1950,export="prank")

# with GSS 1972 data, occopations are coded with 1970 census code:
data("gss1972")
# mapping to 2010 occupation code and export a data frame showing the
# detailed information about code mapping and occ percentile rank
# check data
head(gss1972)
gen_prank1 <- occ_rank(occ = gss1972$occ, cohort = gss1972$birthyr,
                       input_year = 1970,
                       output_year = 2010,export="table")
head(gen_prank1)

# add the generated occupation rank to the input data
gss1972$occ_prank <- occ_rank(occ = gss1972$occ, cohort = gss1972$birthyr,
                              input_year = 1970,
                              output_year = 2010,export="prank")
head(gss1972)

# with GSS 1989 data, occupations are coded with 1980 census code:
data("gss1989")
 # mapping 1980 code to 1990 and export the percentile rank
gen_prank2 <- occ_rank(occ = gss1989$occ80, cohort = gss1989$birthyr,
         input_year = 1980,
         output_year = 1990,export="prank")
head(gen_prank2)

# with GSS 2018 data, occupations are coded with 2010 census code.
# Then the index are directly based on 2010 occupation code
data("gss2018")
gen_prank3 <- occ_rank(occ = gss2018$occ10, cohort = gss2018$birthyr,
        input_year = 2010,
        output_year = 2010,export="table")
head(gen_prank3)
}
