% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_rank_viz.R
\name{occ_rank_plot}
\alias{occ_rank_plot}
\title{Visualizing and exploring the returned occupation percentile rank}
\usage{
occ_rank_plot(occ, cohort, input_year, output_year, plot_occ = NULL)
}
\arguments{
\item{occ}{occupation code in the input year}

\item{cohort}{birth cohort}

\item{input_year}{input year of occupational code(s), e.g., 1950, 1960, 1970, 1980, 1990, 2000, 2010.}

\item{output_year}{The year occupational code(s) is/are crosswalked to, can only be 1950, 1990, and 2010}

\item{plot_occ}{pick up an occupation and plot the percentile ranks in
different birth cohort in the existing data}
}
\value{
occupation index or a data frame including all detailed information
}
\description{
Visualizing and exploring the returned occupation percentile rank
}
\examples{
library(OccRank)
library(haven)

# load data
data("gss1972")
# mapping to 2010 occupation code and visualize the percentile rank for
# selected occupations
# Accountants
occ_rank_plot(occ = gss1972$occ, cohort = gss1972$birthyr,
              input_year = 1970,
              output_year = 2010,plot_occ = 1)

# Shipping and receiving clerks
occ_rank_plot(occ = gss1972$occ, cohort = gss1972$birthyr,
              input_year = 1970,
              output_year = 2010,plot_occ = 374)
# if the occupation code is wrong, there will be an error message
# Do not run:
occ_rank_plot(occ = gss1972$occ, cohort = gss1972$birthyr,
              input_year = 1970,
                output_year = 2010,plot_occ = 800)

}
