% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccRank.R
\docType{data}
\name{year_refocc}
\alias{year_refocc}
\title{Integrated Occupational Codes Crosswalk File}
\format{
A data frame with 17 rows and 4 variables
}
\usage{
data(year_refocc)
}
\description{
An integrated crosswalk table mapping occupational codes from the input year
to the output year (i.e., 1950, 1990, and 2010).
}
\details{
The variables are as follows:

\itemize{
  \item year The input year of occupation code(s).
  \item dict1950 Crosswalk input occupation code(s) to the code(s) in census
  1950.
  \item dict1990 Crosswalk input occupation code(s) to the code(s) in census
  1990.
  \item dict2010 Crosswalk input occupation code(s) to the code(s) in census
  2010.
}
}
\keyword{datasets}
